Move runtime log store into runtimehost service
Runtime log storage primitives now live under the runtimehost service boundary:\n\n    import "github.com/valon-technologies/gestalt/server/services/runtimehost/runtimelogs"\n\n    store := runtimelogs.NewMemoryStore()\n    _ = store.RegisterSession(ctx, runtimelogs.SessionRegistration{\n        RuntimeProviderName: "modal",\n        SessionID:           sessionID,\n    })\n    _, _ = store.AppendSessionLogs(ctx, "modal", sessionID, []runtimelogs.AppendEntry{{\n        Stream:  runtimelogs.StreamRuntime,\n        Message: "runtime booted",\n    }})\n\nThis is an import-path and ownership move only: the Store contract, MemoryStore retention behavior, stream constants, and runtime log host transport behavior are unchanged.
