+++ gestaltd/services/runtimehost/runtimelogs/memory.go
